+++ src/Capsule/UnexpectedValueException.php
@@ -1,4 +1,5 @@
-<?php namespace Atyantik\Capsule;
+<?php
+namespace Atyantik\Capsule;
 
 /**
  * Created by PhpStorm.
